+++ packages/tailwindcss-language-server/src/resolve-css-imports.ts
@@ -1,0 +1,8 @@
+import postcss from 'postcss'
+import postcssImport from 'postcss-import'
+
+const resolveImports = postcss([postcssImport()])
+
+export function resolveCssImports() {
+  return resolveImports
+}
